RFC for cleaning up how we do routing
